AO3-6085 Fix eager loading for zeitwerk mode
Enable config.eager_load in test, but only for CI environments.

Stop using config.autoload_paths, which is not recommended in Rails 6+.

Fix constant paths to match Zeitwerk's expectations:

- Add app/models sub-directories to config.eager_load_paths, because those
  constants are already referenced everywhere.

- The live validations plugin is basically a monkey-patch for ActionView, and
  is not very modular. It's simplest to just move it into
  config/initializers/monkeypatches, where Zeitwerk does not complain
  about the file path and constant path mismatch.

- The commentable plugin is written like a concern, so renaming the constants
  is easy.

Also remove an outdated README (we no longer set up ArchiveConfig in an
initializer) and an unused gem initializer.
