--- config/initializers/archive_config/README
@@ -1,2 +1,0 @@
-Files are read in alphabetical order.
-archive_config must come first.
