+++ config/initializers/gem-plugin_config/acts_as_commentable.rb
@@ -1,5 +1,3 @@
-%w(acts_as_commentable commentable_entity comment_methods).each do |file|
-  require "plugins/acts_as_commentable/#{file}"
-end
+require "acts_as_commentable/commentable"
 
-ActiveRecord::Base.send :include, ActiveRecord::Acts::Commentable
+ActiveRecord::Base.include ActsAsCommentable::Commentable
