--- config/initializers/gem-plugin_config/phusion_passenger_config.rb
@@ -1,9 +1,0 @@
-if defined?(PhusionPassenger)
-  PhusionPassenger.on_event(:starting_worker_process) do |forked|
-    if forked
-      Rails.cache.instance_variable_get(:@data).reset if Rails.cache.class == ActiveSupport::Cache::MemCacheStore
-    else
-      # No need to do anything.
-    end
-  end
-end
